add docs and fix render.sh

+++ render.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 DECAPOD_BASE_URL=https://github.com/openinfradev/decapod-base-yaml.git
-BRANCH="master"
+BRANCH="main"
 
 if [ $# -eq 0 ]; then
   echo 'Error: Missing Arguments "application name"'
@@ -19,7 +19,7 @@ if [ $? -ne 0 ]; then
 fi
 
 echo "Start to build for $APP_NAME"
-
+cp -r decapod-base-yaml/$APP_NAME/base $APP_NAME/
 for i in `ls ${APP_NAME}/site`
 do
 
@@ -27,17 +27,25 @@ do
    echo "[$i] Clean up existing $APP_NAME-manifest.yaml"
    rm -f $OUTPUT_PATH
    mkdir -p $APP_NAME/output/$i/
-   docker rm -f kustomize-build-$APP_NAME
+
    echo "[$i] Rendering $APP_NAME-manifest.yaml for $i site"
-   docker run -it -v $(pwd)/$APP_NAME/site:/site -v $(pwd)/decapod-yaml/$APP_NAME/base:/base -v $(pwd)/$APP_NAME/output:/output --name kustomize-build-$APP_NAME sktdev/decapod-kustomize:alpha-v2.0 /site/$i -o /output/$i/$APP_NAME-manifest.yaml;
+   docker run --rm -i -v $(pwd)/$APP_NAME:/$APP_NAME --name kustomize-build-$APP_NAME sktdev/decapod-kustomize:latest kustomize build --enable_alpha_plugins /$APP_NAME/site/$i -o /$APP_NAME/output/$i/$APP_NAME-manifest.yaml
+   build_result=$?
+
+   if [ $build_result != 0 ]; then
+     exit $build_result
+   fi
 
    if [ -f "$OUTPUT_PATH" ]; then
      echo "[$i] Successfully Completed!"
    else
-     echo "[$i] Failed to render -manifest.yaml"
-     break
+     echo "[$i] Failed to render $APP_NAME-manifest.yaml"
+     rm -rf $APP_NAME/base
+     rm -rf decapod-yaml
+     exit 1
    fi
 
 done
 
-rm -rf decapod-yaml
+rm -rf $APP_NAME/base
+rm -rf decapod-base-yaml
